fix pre-commit

+++ programs/drift/src/controller/liquidation/tests.rs
@@ -574,7 +574,7 @@ pub mod liquidate_perp {
         let mut user_stats = UserStats::default();
         let mut liquidator_stats = UserStats::default();
         let state = State {
-            liquidation_margin_buffer_ratio: MARGIN_PRECISION as u32 / 50,
+            liquidation_margin_buffer_ratio: MARGIN_PRECISION / 50,
             initial_pct_to_liquidate: LIQUIDATION_PCT_PRECISION as u16,
             liquidation_duration: 150,
             ..Default::default()
@@ -725,7 +725,7 @@ pub mod liquidate_perp {
         let mut user_stats = UserStats::default();
         let mut liquidator_stats = UserStats::default();
         let state = State {
-            liquidation_margin_buffer_ratio: MARGIN_PRECISION as u32 / 50,
+            liquidation_margin_buffer_ratio: MARGIN_PRECISION / 50,
             initial_pct_to_liquidate: LIQUIDATION_PCT_PRECISION as u16,
             liquidation_duration: 150,
             ..Default::default()
@@ -1511,7 +1511,7 @@ pub mod liquidate_perp {
         let mut user_stats = UserStats::default();
         let mut liquidator_stats = UserStats::default();
         let state = State {
-            liquidation_margin_buffer_ratio: MARGIN_PRECISION as u32 / 50,
+            liquidation_margin_buffer_ratio: MARGIN_PRECISION / 50,
             initial_pct_to_liquidate: (LIQUIDATION_PCT_PRECISION / 10) as u16,
             liquidation_duration: 150,
             ..Default::default()
@@ -1788,7 +1788,7 @@ pub mod liquidate_perp {
         let mut user_stats = UserStats::default();
         let mut liquidator_stats = UserStats::default();
         let state = State {
-            liquidation_margin_buffer_ratio: MARGIN_PRECISION as u32 / 50,
+            liquidation_margin_buffer_ratio: MARGIN_PRECISION / 50,
             initial_pct_to_liquidate: (LIQUIDATION_PCT_PRECISION / 10) as u16,
             liquidation_duration: 150,
             ..Default::default()
@@ -1913,7 +1913,7 @@ pub mod liquidate_perp {
         let mut user_stats = UserStats::default();
         let mut liquidator_stats = UserStats::default();
         let state = State {
-            liquidation_margin_buffer_ratio: MARGIN_PRECISION as u32 / 50,
+            liquidation_margin_buffer_ratio: MARGIN_PRECISION / 50,
             initial_pct_to_liquidate: (LIQUIDATION_PCT_PRECISION / 10) as u16,
             liquidation_duration: 150,
             ..Default::default()
@@ -2631,7 +2631,7 @@ pub mod liquidate_spot {
         let liquidator_key = Pubkey::default();
 
         let state = State {
-            liquidation_margin_buffer_ratio: MARGIN_PRECISION as u32 / 50,
+            liquidation_margin_buffer_ratio: MARGIN_PRECISION / 50,
             initial_pct_to_liquidate: LIQUIDATION_PCT_PRECISION as u16,
             liquidation_duration: 150,
             ..Default::default()
@@ -2843,7 +2843,7 @@ pub mod liquidate_spot {
         let user_key = Pubkey::default();
         let liquidator_key = Pubkey::default();
         let state = State {
-            liquidation_margin_buffer_ratio: MARGIN_PRECISION as u32 / 50,
+            liquidation_margin_buffer_ratio: MARGIN_PRECISION / 50,
             initial_pct_to_liquidate: LIQUIDATION_PCT_PRECISION as u16,
             liquidation_duration: 150,
             ..Default::default()
@@ -2967,7 +2967,7 @@ pub mod liquidate_spot {
         let user_key = Pubkey::default();
         let liquidator_key = Pubkey::default();
         let state = State {
-            liquidation_margin_buffer_ratio: MARGIN_PRECISION as u32 / 50,
+            liquidation_margin_buffer_ratio: MARGIN_PRECISION / 50,
             initial_pct_to_liquidate: LIQUIDATION_PCT_PRECISION as u16,
             liquidation_duration: 150,
             ..Default::default()
@@ -3093,7 +3093,7 @@ pub mod liquidate_spot {
         let liquidator_key = Pubkey::default();
 
         let state = State {
-            liquidation_margin_buffer_ratio: MARGIN_PRECISION as u32 / 50,
+            liquidation_margin_buffer_ratio: MARGIN_PRECISION / 50,
             initial_pct_to_liquidate: LIQUIDATION_PCT_PRECISION as u16,
             liquidation_duration: 150,
             ..Default::default()
@@ -3223,7 +3223,7 @@ pub mod liquidate_spot {
         let liquidator_key = Pubkey::default();
 
         let state = State {
-            liquidation_margin_buffer_ratio: MARGIN_PRECISION as u32 / 50,
+            liquidation_margin_buffer_ratio: MARGIN_PRECISION / 50,
             initial_pct_to_liquidate: (LIQUIDATION_PCT_PRECISION / 10) as u16,
             liquidation_duration: 150,
             ..Default::default()
@@ -3463,7 +3463,7 @@ pub mod liquidate_spot {
         let user_key = Pubkey::default();
         let liquidator_key = Pubkey::default();
 
-        let liquidation_buffer = MARGIN_PRECISION as u32 / 50;
+        let liquidation_buffer = MARGIN_PRECISION / 50;
         let state = State {
             liquidation_margin_buffer_ratio: liquidation_buffer,
             initial_pct_to_liquidate: LIQUIDATION_PCT_PRECISION as u16,
@@ -3825,7 +3825,7 @@ pub mod liquidate_borrow_for_perp_pnl {
         let user_key = Pubkey::default();
         let liquidator_key = Pubkey::default();
 
-        let liquidation_buffer = MARGIN_PRECISION as u32 / 50;
+        let liquidation_buffer = MARGIN_PRECISION / 50;
         liquidate_borrow_for_perp_pnl(
             0,
             1,
@@ -4471,7 +4471,7 @@ pub mod liquidate_borrow_for_perp_pnl {
         let user_key = Pubkey::default();
         let liquidator_key = Pubkey::default();
 
-        let liquidation_buffer = MARGIN_PRECISION as u32 / 50;
+        let liquidation_buffer = MARGIN_PRECISION / 50;
         liquidate_borrow_for_perp_pnl(
             0,
             1,
@@ -4628,7 +4628,7 @@ pub mod liquidate_borrow_for_perp_pnl {
         let user_key = Pubkey::default();
         let liquidator_key = Pubkey::default();
 
-        let liquidation_buffer = MARGIN_PRECISION as u32 / 50;
+        let liquidation_buffer = MARGIN_PRECISION / 50;
         liquidate_borrow_for_perp_pnl(
             0,
             1,
@@ -5071,7 +5071,7 @@ pub mod liquidate_perp_pnl_for_deposit {
             &mut oracle_map,
             now,
             slot,
-            MARGIN_PRECISION as u32 / 50,
+            MARGIN_PRECISION / 50,
             PERCENTAGE_PRECISION,
             150,
         )
@@ -5679,7 +5679,7 @@ pub mod liquidate_perp_pnl_for_deposit {
             &mut oracle_map,
             now,
             slot,
-            MARGIN_PRECISION as u32 / 50,
+            MARGIN_PRECISION / 50,
             PERCENTAGE_PRECISION,
             150,
         )
@@ -5821,7 +5821,7 @@ pub mod liquidate_perp_pnl_for_deposit {
         let user_key = Pubkey::default();
         let liquidator_key = Pubkey::default();
 
-        let liquidation_buffer = MARGIN_PRECISION as u32 / 50;
+        let liquidation_buffer = MARGIN_PRECISION / 50;
         liquidate_perp_pnl_for_deposit(
             0,
             1,
@@ -6093,7 +6093,7 @@ pub mod liquidate_perp_pnl_for_deposit {
         .is_err());
 
         let state = State {
-            liquidation_margin_buffer_ratio: MARGIN_PRECISION as u32 / 50,
+            liquidation_margin_buffer_ratio: MARGIN_PRECISION / 50,
             initial_pct_to_liquidate: (PERCENTAGE_PRECISION / 10) as u16,
             liquidation_duration: 150,
             ..Default::default()
